feat: add syncAdvancedMode to display different sync screens
