+++ public/app/config/config.js
@@ -44,6 +44,7 @@ const TMP_FOLDER = 'tmp';
 const DEFAULT_LANG = 'en';
 const DEFAULT_DEVELOPER_MODE = false;
 const DEFAULT_GEOLOCATION_ENABLED = false;
+const DEFAULT_SYNC_ADVANCED_MODE = false;
 const DEFAULT_PROTOCOL = 'https';
 const DEFAULT_LOGGING_LEVEL = 'info';
 const AUTHENTICATED = 'authenticated';
@@ -79,4 +80,5 @@ module.exports = {
   DEFAULT_USER,
   AUTHENTICATED,
   ANONYMOUS_USERNAME,
+  DEFAULT_SYNC_ADVANCED_MODE,
 };
