+++ public/electron.js
@@ -41,6 +41,8 @@ const {
   GET_APP_INSTANCE_CHANNEL,
   GET_DEVELOPER_MODE_CHANNEL,
   SET_DEVELOPER_MODE_CHANNEL,
+  GET_SYNC_ADVANCED_MODE_CHANNEL,
+  SET_SYNC_ADVANCED_MODE_CHANNEL,
   GET_GEOLOCATION_ENABLED_CHANNEL,
   SET_GEOLOCATION_ENABLED_CHANNEL,
   GET_DATABASE_CHANNEL,
@@ -84,6 +86,8 @@ const {
   postAppInstanceResource,
   patchAppInstanceResource,
   getAppInstance,
+  setSyncAdvancedMode,
+  getSyncAdvancedMode,
 } = require('./app/listeners');
 const isMac = require('./app/utils/isMac');
 
@@ -391,6 +395,18 @@ app.on('ready', async () => {
 
   // called when setting developer mode
   ipcMain.on(SET_DEVELOPER_MODE_CHANNEL, setDeveloperMode(mainWindow, db));
+
+  // called when setting sync advanced mode
+  ipcMain.on(
+    SET_SYNC_ADVANCED_MODE_CHANNEL,
+    setSyncAdvancedMode(mainWindow, db)
+  );
+
+  // called when getting sync advanced mode
+  ipcMain.on(
+    GET_SYNC_ADVANCED_MODE_CHANNEL,
+    getSyncAdvancedMode(mainWindow, db)
+  );
 
   // called when getting geolocation enabled
   ipcMain.on(
