+++ src/config/channels.js
@@ -45,4 +45,6 @@ module.exports = {
   SIGN_IN_CHANNEL: 'auth:signin',
   SIGN_OUT_CHANNEL: 'auth:signout',
   IS_AUTHENTICATED_CHANNEL: 'auth:authenticated:get',
+  GET_SYNC_ADVANCED_MODE_CHANNEL: 'sync:advanced-mode:get',
+  SET_SYNC_ADVANCED_MODE_CHANNEL: 'sync:advanced-mode:set',
 };
