+++ src/config/constants.js
@@ -65,5 +65,6 @@ export const DIFF_STYLES = {
   },
 };
 
+export const DEFAULT_SYNC_ADVANCED_MODE = false;
 export const SYNC_PHASE_PROPERTIES = ['name', 'image', 'description', 'phases'];
 export const SYNC_ITEM_PROPERTIES = ['name', 'description', 'content'];
