+++ src/types/user.js
@@ -17,3 +17,14 @@ export const GET_GEOLOCATION_ENABLED_SUCCEEDED =
   'GET_GEOLOCATION_ENABLED_SUCCEEDED';
 export const SET_GEOLOCATION_ENABLED_SUCCEEDED =
   'SET_GEOLOCATION_ENABLED_SUCCEEDED';
+
+export const FLAG_GETTING_SYNC_ADVANCED_MODE =
+  'FLAG_GETTING_SYNC_ADVANCED_MODE';
+export const FLAG_SETTING_SYNC_ADVANCED_MODE =
+  'FLAG_SETTING_SYNC_ADVANCED_MODE';
+export const SET_SYNC_ADVANCED_MODE = 'SET_SYNC_ADVANCED_MODE';
+export const GET_SYNC_ADVANCED_MODE = 'GET_SYNC_ADVANCED_MODE';
+export const SET_SYNC_ADVANCED_MODE_SUCCEEDED =
+  'SET_SYNC_ADVANCED_MODE_SUCCEEDED';
+export const GET_SYNC_ADVANCED_MODE_SUCCEEDED =
+  'GET_SYNC_ADVANCED_MODE_SUCCEEDED';
